+++ src/commands/hardis/source/push.ts
@@ -6,7 +6,7 @@ import { wrapDeployCommand } from "../../../common/utils/wrapUtils";
 export default class Push extends SfdxCommand {
   public static readonly description = `sfdx-hardis wrapper for sfdx force:source:push that displays tips to solve deployment errors.
 
-See documentation at https://developer.salesforce.com/docs/atlas.en-us.sfdx_cli_reference.meta/sfdx_cli_reference/cli_reference_force_source.htm#cli_reference_force_source_push
+[See documentation of Salesforce command](https://developer.salesforce.com/docs/atlas.en-us.sfdx_cli_reference.meta/sfdx_cli_reference/cli_reference_force_source.htm#cli_reference_force_source_push)
 `;
   protected static readonly flagsConfig: FlagsConfig = {
     forceoverwrite: flags.boolean({